requirements.txt
appdirs==1.4.4
cfgv==3.2.0
DateTime==4.3
distlib==0.3.1
filelock==3.0.12
identify==1.5.13
nodeenv==1.5.0
numpy==1.20.1
pandas==1.2.2
pre-commit==2.10.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
toml==0.10.2
virtualenv==20.4.2
zope.interface==5.2.0
